ref: make request class more dry
