+++ app/Http/Requests/Post/DeletePostRequest.php
@@ -3,14 +3,13 @@
 namespace App\Http\Requests\Post;
 
 use App\Http\Requests\BaseRequest;
+use App\Http\Requests\Traits\PostFromRouteTrait;
 use App\Models\Post\Post;
-use App\Models\User\User;
 use App\Port\Core\Post\DeletePostPort;
 
 class DeletePostRequest extends BaseRequest implements DeletePostPort
 {
-    protected ?User $authenticatedUser = null;
-    protected ?Post $postFromRoute = null;
+    use PostFromRouteTrait;
 
     /**
      * Determine if the user is authorized to make this request.
@@ -33,23 +32,5 @@ class DeletePostRequest extends BaseRequest implements DeletePostPort
     public function getPost(): Post
     {
         return $this->getPostFromRoute();
-    }
-
-    protected function getPostFromRoute(): Post
-    {
-        if (is_null($this->postFromRoute)) {
-            $this->postFromRoute = $this->route('postID');
-        }
-
-        return $this->postFromRoute;
-    }
-
-    protected function getAuthenticatedUser(): User
-    {
-        if (is_null($this->authenticatedUser)) {
-            $this->authenticatedUser = $this->getUserOrFail();
-        }
-
-        return $this->authenticatedUser;
     }
 }
